Move initialize_placeholder_tokens_from_initializer_token(...) to _shared/ directory.
