+++ src/invoke_training/training/pipelines/stable_diffusion/textual_inversion_sd.py
@@ -25,7 +25,10 @@ from invoke_training.training._shared.data.data_loaders.textual_inversion_sd_dat
 )
 from invoke_training.training._shared.optimizer.optimizer_utils import initialize_optimizer
 from invoke_training.training._shared.stable_diffusion.model_loading_utils import load_models_sd
-from invoke_training.training._shared.stable_diffusion.textual_inversion import add_tokens_to_tokenizer
+from invoke_training.training._shared.stable_diffusion.textual_inversion import (
+    add_tokens_to_tokenizer,
+    initialize_placeholder_tokens_from_initializer_token,
+)
 from invoke_training.training.pipelines.stable_diffusion.finetune_lora_sd import (
     cache_vae_outputs,
     generate_validation_images,
@@ -60,32 +63,6 @@ def _save_ti_embeddings(
     learned_embeds_dict = {placeholder_token: learned_embeds.detach().cpu()}
 
     save_state_dict(learned_embeds_dict, save_path)
-
-
-def initialize_placeholder_tokens_from_initializer_token(
-    tokenizer: CLIPTokenizer, text_encoder: CLIPTextModel, initializer_token: str, placeholder_tokens: list[str]
-) -> list[int]:
-    # Convert the initializer_token and placeholder_token to token ids.
-    initializer_token_ids = tokenizer.encode(initializer_token, add_special_tokens=False)
-    if len(initializer_token_ids) > 1:
-        raise ValueError(
-            f"The initializer_token '{initializer_token}' gets tokenized to {len(initializer_token_ids)} tokens."
-            " Choose a different initializer that maps to a single token."
-        )
-    initializer_token_id = initializer_token_ids[0]
-    placeholder_token_ids = tokenizer.convert_tokens_to_ids(placeholder_tokens)
-
-    # convert_tokens_to_ids returns a `int | list[int]` type, but since we pass in a list it should always return a
-    # list.
-    assert isinstance(placeholder_token_ids, list)
-
-    # Initialize the newly-added placeholder token(s) with the embeddings of the initializer token.
-    token_embeds = text_encoder.get_input_embeddings().weight.data
-    with torch.no_grad():
-        for token_id in placeholder_token_ids:
-            token_embeds[token_id] = token_embeds[initializer_token_id].clone()
-
-    return placeholder_token_ids
 
 
 def _initialize_placeholder_tokens_from_initial_embedding(
